load dynamic components
