+++ core/new-gui/package.json
@@ -58,6 +58,7 @@
     "lodash": "~4.17.21",
     "lodash-es": "~4.17.14",
     "mapbox-gl": "^2.2.0",
+    "ng-dynamic-component": "^6.0.1",
     "ng-gapi": "0.0.94",
     "ng-zorro-antd": "^10.0.0",
     "ng2-file-upload": "^1.4.0",
