+++ core/new-gui/src/app/app.module.ts
@@ -1,5 +1,5 @@
 import { registerLocaleData } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import en from '@angular/common/locales/en';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -76,7 +76,7 @@ import { ProductTourComponent } from './workspace/component/product-tour/product
 import { PropertyEditorComponent } from './workspace/component/property-editor/property-editor.component';
 import { TypeCastingDisplayComponent } from './workspace/component/property-editor/typecasting-display/type-casting-display.component';
 import { ResultPanelToggleComponent } from './workspace/component/result-panel-toggle/result-panel-toggle.component';
-import { ResultPanelComponent, RowModalComponent } from './workspace/component/result-panel/result-panel.component';
+import { ResultPanelComponent } from './workspace/component/result-panel/result-panel.component';
 import { VisualizationPanelContentComponent } from './workspace/component/visualization-panel-content/visualization-panel-content.component';
 import { VisualizationPanelComponent } from './workspace/component/visualization-panel/visualization-panel.component';
 import { WorkflowEditorComponent } from './workspace/component/workflow-editor/workflow-editor.component';
@@ -89,8 +89,11 @@ import { NzCardModule } from 'ng-zorro-antd/card';
 import { NzStatisticModule } from 'ng-zorro-antd/statistic';
 import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzAvatarModule } from 'ng-zorro-antd/avatar';
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { BlobErrorHttpInterceptor } from './common/service/blob-error-http-interceptor.service';
+import { ConsoleFrameComponent } from './workspace/component/result-panel/console-frame/console-frame.component';
+import { ResultTableFrameComponent } from './workspace/component/result-panel/result-table-frame/result-table-frame.component';
+import { DynamicModule } from 'ng-dynamic-component';
+import { RowModalComponent } from './workspace/component/result-panel/result-panel-modal.component';
 
 registerLocaleData(en);
 
@@ -137,7 +140,9 @@ registerLocaleData(en);
     TypeCastingDisplayComponent,
     ResultDownloadComponent,
     NgbdModalWorkflowShareAccessComponent,
-    NgbdModalUserFileShareAccessComponent
+    NgbdModalUserFileShareAccessComponent,
+    ConsoleFrameComponent,
+    ResultTableFrameComponent
   ],
   imports: [
     BrowserModule,
@@ -189,7 +194,8 @@ registerLocaleData(en);
     NzCardModule,
     NzStatisticModule,
     NzTagModule,
-    NzAvatarModule
+    NzAvatarModule,
+    DynamicModule
   ],
   entryComponents: [
     NgbdModalAddWorkflowComponent,
