+++ core/new-gui/src/app/workspace/component/result-panel/result-panel.component.html
@@ -1,58 +1,3 @@
 <div class="texera-workspace-result-panel-body texera-workflow-component-body" [hidden]="!showResultPanel">
-
-  <div class="texera-panel-message" *ngIf="errorMessages !== undefined">
-    <div *ngFor="let item of errorMessages | keyvalue">
-      {{item.key}}:{{item.value}}
-    </div>
-  </div>
-  <!-- Use [hidden] instead of *ngIf because in *ngIf, when the predicate returns
-    false, the element inside the *ngIf clause will not exist, which @ViewChild(MatPaginator)
-    will not be able to get the paginator when the correct execution result is returned
-    for the first time, since <mat-paginator></mat-paginator> will still be undefined as
-    the elements in *ngIf is still loading. [hidden] attribute solves this issue by not
-    removing the elements, but rather modify their css to make these elements invisible
-    to the users-->
-  <div [hidden]="chartType">
-    <button class="btn btn-secondary" *ngIf="breakpointTriggerInfo !== undefined" (click)="onClickSkipTuples()"
-      [disabled]="!breakpointAction">
-      Skip Records
-    </button>
-    <div class="result-table" [hidden]="!currentColumns">
-      <nz-table
-        #basicTable
-        [nzData]="currentResult"
-        nzBordered="true"
-        [nzFrontPagination]="isFrontPagination"
-        [nzLoading]="isLoadingResult"
-        [nzTotal]="totalNumTuples"
-        [nzSize]="'small'"
-        [nzScroll]="{y: '200px'}"
-        [nzTableLayout]="'fixed'"
-        [nzPageIndex]="currentPageIndex"
-        [nzPageSize]="pageSize"
-        [nzPaginationPosition]="'bottom'"
-        (nzQueryParams)="onTableQueryParamsChange($event)"
-      >
-        <thead>
-          <tr>
-            <th *ngFor="let column of currentColumns">
-              {{column.header}}
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let row of basicTable.data; let i=index;">
-            <td nzEllipsis (click)="open(row)" *ngFor="let column of currentColumns" class="table-row">
-              {{column.getCell(row)}}
-            </td>
-          </tr>
-        </tbody>
-      </nz-table>
-    </div>
-  </div>
-
-  <div>
-    <texera-visualization-panel [operatorID]="resultPanelOperatorID"></texera-visualization-panel>
-  </div>
-
+  <ndc-dynamic [ndcDynamicComponent]="component"></ndc-dynamic>
 </div>
