+++ core/new-gui/src/app/workspace/component/visualization-panel/visualization-panel.component.html
@@ -1,2 +1,2 @@
-<button  *ngIf="displayVisualizationPanel" nz-button (click)="onClickVisualize()">Visualize</button>
+<button nz-button (click)="onClickVisualize()">Visualize</button>
