+++ core/new-gui/src/app/workspace/component/visualization-panel/visualization-panel.component.spec.ts
@@ -69,7 +69,7 @@ describe('VisualizationPanelComponent', () => {
     component.ngOnChanges();
     fixture.detectChanges();
 
-    console.log(component.displayVisualizationPanel);
+
 
     const element: HTMLElement = fixture.nativeElement;
     const button = element.querySelector('button');
